+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-CustomTemplates-client/src/test/java/com/graphql_java_generator/samples/basic/client/ValidateCustomQueryIT.java
@@ -17,23 +17,20 @@ import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
 import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.Board;
 import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.Query;
 import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.QueryExecutor;
-import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.SubscriptionExecutor;
+import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.Subscription;
 
 @SpringBootTest()
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = ValidateCustomQueryConfiguration.class)
+@ContextConfiguration(classes = SpringConfiguration.class)
 class ValidateCustomQueryIT {
 
 	@Autowired
 	QueryExecutor query;
 
-	@Autowired
-	SubscriptionExecutor subscription;
-
 	@Test
 	void test_customTemplateInTheProject() throws GraphQLRequestPreparationException, GraphQLRequestExecutionException {
 		// Let's check that our QueryType is generated from the custom template
-		assertTrue(subscription.thisIsADummyFieldToCheckThatThisTemplateIsUsed);
+		assertTrue(new Subscription().thisIsADummyFieldToCheckThatThisTemplateIsUsed);
 	}
 
 	@Test