+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-StarWars-client/src/test/java/com/graphql_java_generator/mavenplugin/samples/simple/client/subscription/CharacterSubscriptionCallback.java
@@ -3,6 +3,8 @@
  */
 package com.graphql_java_generator.mavenplugin.samples.simple.client.subscription;
 
+import java.util.concurrent.CountDownLatch;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -25,6 +27,9 @@ public class CharacterSubscriptionCallback implements SubscriptionCallback<Chara
 	String lastReceivedClose = null;
 	Throwable lastReceivedError = null;
 
+	/** A latch that will be freed when a the first notification arrives for this subscription */
+	public CountDownLatch latchForMessageReception = new CountDownLatch(1);
+
 	@Override
 	public void onConnect() {
 		connected = true;
@@ -34,7 +39,7 @@ public class CharacterSubscriptionCallback implements SubscriptionCallback<Chara
 	public void onMessage(Character t) {
 		lastReceivedMessage = t;
 		logger.debug("Received {} {}", t.getClass().getSimpleName(), t);
-		SubscriptionIT.currentThread.interrupt();
+		latchForMessageReception.countDown();
 	}
 
 	@Override